refactor: modularize test suite with proper separation of concerns
- Create tests/test_ContentType.py for ContentType plugin tests
  - Move ContentType plugin functionality tests from test_cli.py
  - Add CLI integration tests for ContentType environment variable behavior
  - Include comprehensive test coverage for content type detection and processing

- Refactor tests/test_cli.py to focus on CLI functionality only
  - Remove redundant EntityReference plugin tests (already in test_EntityReference.py)
  - Remove ContentType plugin tests (moved to dedicated file)
  - Improve test robustness by removing hardcoded plugin dependencies
  - Add comprehensive CLI error handling tests
  - Make plugin help tests dynamic and generic
  - Remove unused imports (tempfile)

- Enhance test maintainability and isolation
  - Tests no longer break when plugin names change
  - Better separation between unit tests and integration tests
  - Improved error path coverage for CLI functionality
  - More robust plugin discovery testing

This refactoring ensures each test file has a clear, focused responsibility
and follows the project's naming conventions for plugin test files.
